Add icon_svg option to company phone, email and address shortcodes
Add check for view before trying to load it in single.blade.php
Fix issue with title fields turning into an Array
Fix bug when you have more than one set of radio fields at the same time

+++ src/helpers.php
@@ -325,10 +325,10 @@ if (!function_exists('get_slider_options_v2')) {
       // Initialize only if needed
       $option = [];
       if (isset($slidesPerView[$breakpoint]) && !empty($slidesPerView[$breakpoint])) {
-        $option['slidesPerView'] = (int) $slidesPerView[$breakpoint];
+        $option['slidesPerView'] = (float) $slidesPerView[$breakpoint];
       }
       if (isset($spaceBetween[$breakpoint]) && !empty($spaceBetween[$breakpoint])) {
-        $option['spaceBetween'] = (int) $spaceBetween[$breakpoint];
+        $option['spaceBetween'] = (float) $spaceBetween[$breakpoint];
       }
       if ($option) {
         $slider_options[$value] = $option;
